refactor(codemirror): extract inline decoration effects
